code is now cleaner with functions in seperate files, added reset option to start new game
